+++ internal/metrics/helpers_test.go
@@ -4,6 +4,8 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
+	"go.opencensus.io/metric/metricdata"
 	"go.opencensus.io/stats/view"
 )
 
@@ -17,11 +19,41 @@ func testDataRetrieval(v *view.View, t *testing.T, want string) {
 	if err != nil {
 		t.Fatalf("%s: failed to retrieve data line %s", name, err)
 	}
-	if len(data) != 1 {
+
+	if want != "" && len(data) != 1 {
 		t.Fatalf("%s: received incorrect number of data rows: %d", name, len(data))
 	}
+	if want == "" && len(data) > 0 {
+		t.Fatalf("%s: received incorrect number of data rows: %d", name, len(data))
+	} else if want == "" {
+		return
+	}
 
-	if !strings.HasPrefix(data[0].String(), want) {
-		t.Errorf("%s: Found unexpected data row: \nwant: %s\ngot: %s\n", name, want, data[0].String())
+	dataString := data[0].String()
+
+	if want != "" && !strings.HasPrefix(dataString, want) {
+		t.Errorf("%s: Found unexpected data row: \nwant: %s\ngot: %s\n", name, want, dataString)
+	}
+}
+
+func testMetricRetrieval(metrics []*metricdata.Metric, t *testing.T, labels []metricdata.LabelValue, value int64, name string) {
+	found := false
+	for _, metric := range metrics {
+		if metric.Descriptor.Name != name {
+			found = true
+			continue
+		}
+		gotLabels := metric.TimeSeries[0].LabelValues
+		gotValue := metric.TimeSeries[0].Points[0].Value
+
+		if diff := cmp.Diff(gotLabels, labels); diff != "" {
+			t.Errorf("Failed to find metric labels:\n%s", diff)
+		}
+		if diff := cmp.Diff(gotValue, value); diff != "" {
+			t.Errorf("Failed to find metric value:\n%s", diff)
+		}
+	}
+	if !found {
+		t.Errorf("Could not find metric %s", name)
 	}
 }